LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/drawing-text-to-the-canvas/2d.text.drawing.style.spacing-expected.txt

2d.text.drawing.style.spacing
Testing letter spacing and word spacing
Actual output:

FAIL Testing letter spacing and word spacing assert_equals: ctx.letterSpacing === '0px' (got [undefined], expected 0px[string]) expected (string) "0px" but got (undefined) undefined
